feat: WalletConnect v2 connector (#1500)
* feat: implements v2 connector

* chore: packages

* refactor: example app params

* chore: yarn lock

* refactor: proxies all requests through ethereum-provider

* refactor: rm providers

* refactor: cleanup

* refactor: reimplements `walletconnect` v1 & v2 providers to work simultaneously based on init opts

* chore: reverts demo app walletconnect module init

* Restore yarn.lock from develop and bump wc verison in demo for testing

* Update readme types, version bump and version to int

* Add metaData types and pass through

* Update docs for WC

* Bump packages to the latest

* refactor: updates `ethereum-provider` to handle optionalChains

* Revert yarn.lock, yarn-it and remove logs

* Add WC explorer to docs for WC v2

* Revert api key change

---------
