+++ packages/demo/src/App.svelte
@@ -104,7 +104,12 @@
   const coinbaseWallet = coinbaseModule()
 
   const walletConnect = walletConnectModule({
-    connectFirstChainId: true
+    connectFirstChainId: true,
+    version: 2,
+    projectId: 'f6bd6e2911b56f5ac3bc8b2d0e2d7ad5',
+    qrcodeModalOptions: {
+    mobileLinks: ['rainbow', 'metamask', 'argent', 'trust', 'imtoken', 'pillar']
+    }
   })
   const portis = portisModule({
     apiKey: 'b2b7586f-2b1e-4c30-a7fb-c2d1533b153b'
