+++ packages/walletconnect/README.md
@@ -1,10 +1,13 @@
 # @web3-onboard/walletconnect
 
-## Wallet module for connecting Wallet Connect wallets to web3-onboard
+## Wallet module for connecting Wallet Connect to web3-onboard
 
 ### Install
 
 `npm i @web3-onboard/core @web3-onboard/walletconnect`
+
+## Not all Wallets support WalletConnect V2 currently
+_For an up to date list please see the [WalletConnect Explorer](https://explorer.walletconnect.com/?version=2)_
 
 ## Options
 
@@ -15,7 +18,25 @@ type WalletConnectOptions = {
     mobileLinks: string[] // set the order and list of mobile linking wallets
   }
   connectFirstChainId?: boolean // if true, connects to the first network chain provided
-}
+} & (
+  | {
+      /**
+       * Defaults to version: 1 - this behavior will be deprecated after the WalletConnect v1 sunset
+       */
+      version?: 1
+    }
+  | {
+      /**
+       * Project ID associated with [WalletConnect account](https://cloud.walletconnect.com)
+       */
+      projectId: string
+
+      /**
+       * Defaults to version: 1 - this behavior will be deprecated after the WalletConnect v1 sunset
+       */
+      version: 2
+    }
+)
 ```
 
 ## Usage
@@ -24,16 +45,28 @@ type WalletConnectOptions = {
 import Onboard from '@web3-onboard/core'
 import walletConnectModule from '@web3-onboard/walletconnect'
 
-// initialize the module with options
-const walletConnect = walletConnectModule({
+const wcV1InitOptions = {
   bridge: 'YOUR_CUSTOM_BRIDGE_SERVER',
   qrcodeModalOptions: {
-    mobileLinks: ['rainbow', 'metamask', 'argent', 'trust', 'imtoken', 'pillar']
+    mobileLinks: ['metamask', 'argent', 'trust',]
   },
   connectFirstChainId: true
-})
+}
+
+const wcV2InitOptions = {
+  version: 2,
+  /**
+   * Project ID associated with [WalletConnect account](https://cloud.walletconnect.com)
+   */
+  projectId: 'abc123...'
+}
+
+// initialize the module with options
+// If version isn't set it will default to V1 until V1 sunset
+const walletConnect = walletConnectModule(wcV2InitOptions || wcV1InitOptions)
 
 // can also initialize with no options...
+// Defaults to V1 until V1 sunset
 // const walletConnect = walletConnectModule()
 
 const onboard = Onboard({
